Airport-Management-System-master/requirements.txt

Babel
Brotli
cffi
cssselect2
Django
django-admin-interface
django-autoslug
django-colorfield
django-countries
django-flat-responsive
django-flat-theme
django-money
django-phonenumber-field
django-weasyprint
django-widget-tweaks
fonttools
html5lib
phonenumbers
Pillow
py-moneyed
pycparser
pydyf
pymemcache
pyphen
python-decouple
pytz
six
sqlparse
tinycss2
weasyprint
webencodings
zopfli
